backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
Pillow==9.4.0
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.10.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
